TemporalNetworks

TemporalNetworks.DegreeNormalization

TemporalNetworks.IdentityNormalization

TemporalNetworks.Multiplex

TemporalNetworks.NonMultiplexCompressed

TemporalNetworks.RayleighBalancing

TemporalNetworks.SpatTempMatching

TemporalNetworks.SpatTempMatchingNonMultiplex

TemporalNetworks.bipartite_minimum_weight_edge_cover

TemporalNetworks.bisection

TemporalNetworks.DegreeNormalization — Type

(:: DegreeNormalization)(L :: Matrix)

Normalizes Laplacian with resepct to the standard normalization $D^{-1/2}LD^{-1/2}$ where D is diagonal with i-th entry equal to deg(i).

TemporalNetworks.IdentityNormalization — Type

(:: IdentityNormalization)(L :: Matrix)

Trivial Normalization of the Laplacian. Simply returns L.

TemporalNetworks.Multiplex — Type

(:: Multiplex)(mlgraph :: MultilayerGraph)

Connection type :: TemporalConnectivity for MultilayerGraph. Returns W_spat, W_temp such that $Wˢᵖᵃᵗ = ⨁ᵢWⁱ$ and $Wᵗᵉᵐᵖ = W'⊗Iₙ$.

Equivalently,

W_spat = directsum(mlgraph.W) where W :: Vector{Matrix}

W_temp = kron(Wt, Matrix{Float64}(I, mlgraph.N, mlgraph.N)) where Wt is a T x T matrix describing how the layers are connected (symmetric supradiagonal).

TemporalNetworks.NonMultiplexCompressed — Type

(:: NonMultiplexCompressed)(mlgraph :: MultilayerGraph)

Nonmultiplex connection type :: TemporalConnectivity for MultilayerGraph. Returns W_spat, W_temp such that:

W_spat = directsum(mlgraph.W) where W :: Vector{Matrix}

W_temp[i,j] ≂̸ 0 iff

mlgraph.W[k][q,:] ≂̸ 0

mlgraph.W[l][r,:] ≂̸ 0

where $(k,q) ∈ Rᴺ × Rᵀ$ and $(l,r) ∈ Rᴺ × Rᵀ$ are the separated spacetime indices corresponding to $i ∈ Rᴺᵀ$ and $j ∈ Rᴺᵀ$ respectively.

TemporalNetworks.RayleighBalancing — Type

RayleighBalancing(ind :: Int64)

Creates DiffusionEstimator instance to compute the diffusion constant $a$ for the supra-Laplacian arising from a non-multiplex network by matching the spatial and temproral Rayleigh quotient contributions with respect to the eigenvalue indexed by ind.

Recommendeded for nonmultiplex networks.

(rb:: RayleighBalancing)(mlgraph :: MultilayerGraph ,norm :: Normalization, L_spat :: Matrix{Float64}, L_temp :: Matrix{Float64})

Computes the diffusion constant $a$ using the Rayleigh balancing criterion. See [FroylandKaliaKoltai2024] for details on this heuristic.

TemporalNetworks.SpatTempMatching — Type

(:: SpatTempMatching)(::SpatTempMatching)(mlgraph :: MultilayerGraph{T}, norm :: Normalization, L_spat :: Matrix{Float64}, L_temp :: Matrix{Float64}) where T <: Multiplex

Computes the diffusion constant $a$ by matching the second spatial eigenvalue (first non-trivial spatial eigenvalue) to the first temporal eigenvalue for the supra-Laplacian corresponding to a multiplex temporal network.

The multiplex structure guarantees a unique smallest intersection point. See [FroylandKoltai2023] and [AtnipFroylandKoltai2024] for details on this heuristic.

TemporalNetworks.SpatTempMatchingNonMultiplex — Type

SpatTempMatchingNonMultiplex(evec_ind :: Int64, temp_ind :: Union{Int64, Nothng}, thresh :: Float64)

Creates DiffusionEstimator instance to compute the diffusion constant $a$ for the supra-Laplacian arising from a non-multiplex network by matching the eigenvalue corresponding to evec_ind to the corresponding multiplex temporal eigenvalue with index temp_ind.

The parameter thresh is used by isspat(::MultilayerGraph{NonMultiplex}, ...) to check for spatial-like behaviour.

(:: SpatTempMatchingNonMultiplex)(::SpatTempMatching)(mlgraph :: MultilayerGraph{T}, norm :: Normalization, L_spat :: Matrix{Float64}, L_temp :: Matrix{Float64}) where T <: NonMultiplex

Computes the diffusion constant $a$ using the matching heuristic for non-multiplex networks. See [FroylandKaliaKoltai2024] for details on this heuristic.

TemporalNetworks.bipartite_minimum_weight_edge_cover — Method

Computes the minimum-weight edge cover of a bipartite graph G. We assume there are m nodes on the left and n nodes on the right, where m≥n. The (i,j)th entry of the m×n matrix G contains the weight of the edge joining node i and node j. A binary m×n array is output, encoding the cover.

TemporalNetworks.bisection — Method

bisection(f, a, b; fa = f(a), fb = f(b), ftol, wtol)

Bisection algorithm for finding the root $f(x) ≈ 0$ within the initial bracket [a,b].

Returns a named tuple (x = x, fx = f(x), isroot = ::Bool, iter = ::Int, ismaxiter = ::Bool).

Terminates when either

abs(f(x)) < ftol (isroot = true),

the width of the bracket is ≤wtol (isroot = true, to account for discontinuities),

maxiter number of iterations is reached. (isroot = false, maxiter = true).

which are tested for in the above order. Therefore, care should be taken not to make wtol too large.